refactor: search 도메인 패키지 정리, 쿼리 빌딩 팩토리 패턴 이용해서 리팩토링
