+++ src/main/kotlin/mashup/spring/elegant/search/domain/search/QueryBuilderExtension.kt
@@ -1,13 +1,13 @@
 package mashup.spring.elegant.search.domain.search
 
-import mashup.spring.elegant.search.service.BasicSearchService
+import mashup.spring.elegant.search.domain.search.boost.LOCATION_BOOST
+import mashup.spring.elegant.search.domain.search.boost.LOCATION_LIMIT
+import mashup.spring.elegant.search.domain.search.enums.ShopField
 import mashup.spring.elegant.search.util.translateDay
 import org.elasticsearch.index.query.BoolQueryBuilder
 import org.elasticsearch.index.query.QueryBuilder
 import org.elasticsearch.index.query.QueryBuilders
-import org.elasticsearch.index.query.functionscore.FunctionScoreQueryBuilder
 import org.joda.time.LocalDateTime
-import org.springframework.beans.factory.annotation.Value
 import org.springframework.data.domain.PageRequest
 import org.springframework.data.elasticsearch.core.query.NativeSearchQueryBuilder
 import java.time.LocalDate
