--- src/main/kotlin/mashup/spring/elegant/search/domain/search/SearchType.kt
@@ -1,5 +1,0 @@
-package mashup.spring.elegant.search.domain.search
-
-enum class SearchType {
-    KEYWORD, CATEGORY
-}
