+++ src/main/kotlin/mashup/spring/elegant/search/domain/search/boost/BoostFunctions.kt
@@ -1,5 +1,6 @@
-package mashup.spring.elegant.search.domain.search
+package mashup.spring.elegant.search.domain.search.boost
 
+import mashup.spring.elegant.search.domain.search.enums.ShopField
 import org.elasticsearch.common.lucene.search.function.FieldValueFactorFunction
 import org.elasticsearch.index.query.QueryBuilders
 import org.elasticsearch.index.query.functionscore.FieldValueFactorFunctionBuilder
@@ -14,11 +15,13 @@ fun getReviewBooster()  = arrayOf(
     ),
     FunctionScoreQueryBuilder.FilterFunctionBuilder(
         ScoreFunctionBuilders
-            .linearDecayFunction(ShopField.REVIEW_AVG.field,
-                                 REVIEW_AVG_ORIGIN,
-                                 REVIEW_AVG_SCALE,
-                                 REVIEW_AVG_OFFSET,
-                                 REVIEW_AVG_DECAY)
+            .linearDecayFunction(
+                ShopField.REVIEW_AVG.field,
+                REVIEW_AVG_ORIGIN,
+                REVIEW_AVG_SCALE,
+                REVIEW_AVG_OFFSET,
+                REVIEW_AVG_DECAY
+            )
     )
 )
 
@@ -26,10 +29,12 @@ fun getNewShopBooster () = arrayOf(
     FunctionScoreQueryBuilder.FilterFunctionBuilder(
         QueryBuilders.rangeQuery(ShopField.CREATED_DATE.field).gte(NEW_FILTER_FROM).lte(NEW_FILTER_TO),
         ScoreFunctionBuilders
-            .gaussDecayFunction(ShopField.CREATED_DATE.field,
-                                NEW_SCORE_ORIGIN,
-                                NEW_SCORE_SCALE,
-                                NEW_SCORE_OFFSET,
-                                NEW_SCORE_DECAY)
+            .gaussDecayFunction(
+                ShopField.CREATED_DATE.field,
+                NEW_SCORE_ORIGIN,
+                NEW_SCORE_SCALE,
+                NEW_SCORE_OFFSET,
+                NEW_SCORE_DECAY
+            )
     )
 )
