+++ src/main/kotlin/mashup/spring/elegant/search/domain/search/enums/SearchType.kt
@@ -1,0 +1,5 @@
+package mashup.spring.elegant.search.domain.search.enums
+
+enum class SearchType {
+    KEYWORD, CATEGORY
+}
