+++ src/main/kotlin/mashup/spring/elegant/search/domain/search/enums/ShopField.kt
@@ -1,4 +1,4 @@
-package mashup.spring.elegant.search.domain.search
+package mashup.spring.elegant.search.domain.search.enums
 
 enum class ShopField (
     val field : String
@@ -10,8 +10,8 @@ enum class ShopField (
     REVIEW_AVG("review.avg"), // type: Float
     REVIEW_COUNT("review.count"), // type: Int
     CREATED_DATE("created_date"), // type: String(yyyy-mm-dd)
-    MENU_NAME("menu.name"), // type: String
-    MENU_CONTENT("menu.content"), // type: String
+    MENU_NAME("menu.name"), // type: String -> Nested
+    MENU_CONTENT("menu.content"), // type: String -> Nested
     DELIVERY_TIME("delivery_time"), //빼도 될것같은데?
     DELIVERY_TIP("delivery_tip"), //빼도 될것같은데?
     DELIVERY_AREA("delivery_area"), // type: String
@@ -19,7 +19,9 @@ enum class ShopField (
     LOCATION("location"), // type: GeoPoint (lat:Double, lon:Double)
     OPEN_HOUR_WEEK("open_hour.week"), // type: String
     OPEN_HOUR_HOUR("open_hour.hour"), // type: Range
-    TAKE_OUT("take_out"); // type: Boolean
+    TAKE_OUT("feature.take_out"), // type: Boolean
+    ONLY_ONE("feature.only_one"), // type: Boolean
+    DISASTER("feature.disaster") // type: Boolean
 
 
 }
