+++ src/main/kotlin/mashup/spring/elegant/search/domain/search/factory/FunctionQueryFactory.kt
@@ -1,6 +1,6 @@
 package mashup.spring.elegant.search.domain.search.factory
 
-import mashup.spring.elegant.search.domain.search.SearchType
+import mashup.spring.elegant.search.domain.search.enums.SearchType
 import org.elasticsearch.index.query.BoolQueryBuilder
 import org.elasticsearch.index.query.functionscore.FunctionScoreQueryBuilder
