+++ src/main/kotlin/mashup/spring/elegant/search/domain/search/factory/impl/BasicFunctionQueryFactory.kt
@@ -1,9 +1,9 @@
 package mashup.spring.elegant.search.domain.search.factory.impl
 
-import mashup.spring.elegant.search.domain.search.SearchType
+import mashup.spring.elegant.search.domain.search.enums.SearchType
 import mashup.spring.elegant.search.domain.search.factory.FunctionQueryFactory
-import mashup.spring.elegant.search.domain.search.getNewShopBooster
-import mashup.spring.elegant.search.domain.search.getReviewBooster
+import mashup.spring.elegant.search.domain.search.boost.getNewShopBooster
+import mashup.spring.elegant.search.domain.search.boost.getReviewBooster
 import org.elasticsearch.common.lucene.search.function.FunctionScoreQuery.*
 import org.elasticsearch.index.query.BoolQueryBuilder
 import org.elasticsearch.index.query.QueryBuilders
@@ -14,7 +14,7 @@ import org.springframework.stereotype.Component
 @Component
 class BasicFunctionQueryFactory :FunctionQueryFactory {
 
-    private val reviewFunctions by lazy {getReviewBooster()}
+    private val reviewFunctions by lazy { getReviewBooster() }
     private val newShopFunctions by lazy { getNewShopBooster() }
 
     override fun create(type: SearchType, query: BoolQueryBuilder): FunctionScoreQueryBuilder
