+++ src/main/kotlin/mashup/spring/elegant/search/domain/search/factory/impl/BasicQueryBaseFactory.kt
@@ -1,6 +1,11 @@
 package mashup.spring.elegant.search.domain.search.factory.impl
 
-import mashup.spring.elegant.search.domain.search.*
+import mashup.spring.elegant.search.domain.search.boost.CATEGORY_BOOST
+import mashup.spring.elegant.search.domain.search.boost.MENU_CONTENT_BOOST
+import mashup.spring.elegant.search.domain.search.boost.MENU_NAME_BOOST
+import mashup.spring.elegant.search.domain.search.boost.SHOP_NAME_BOOST
+import mashup.spring.elegant.search.domain.search.enums.SearchType
+import mashup.spring.elegant.search.domain.search.enums.ShopField
 import mashup.spring.elegant.search.domain.search.factory.QueryBaseFactory
 import org.elasticsearch.index.query.BoolQueryBuilder
 import org.elasticsearch.index.query.QueryBuilders
@@ -19,6 +24,7 @@ class BasicQueryBaseFactory : QueryBaseFactory {
                 QueryBuilders.matchQuery(ShopField.MENU_NAME.field, term).boost(MENU_NAME_BOOST))
             .should(
                 QueryBuilders.matchQuery(ShopField.MENU_CONTENT.field, term).boost(MENU_CONTENT_BOOST))
+
 
         SearchType.CATEGORY -> QueryBuilders.boolQuery()
             .should(
