+++ src/main/kotlin/mashup/spring/elegant/search/domain/search/mapper/ResultMapper.kt
@@ -1,5 +1,6 @@
-package mashup.spring.elegant.search.domain.search
+package mashup.spring.elegant.search.domain.search.mapper
 
+import mashup.spring.elegant.search.domain.search.Shop
 import mashup.spring.elegant.search.dto.SearchDto
 import org.springframework.data.elasticsearch.core.SearchHits
