--- src/main/kotlin/mashup/spring/elegant/search/dto/Feature.kt
@@ -1,10 +1,0 @@
-package mashup.spring.elegant.search.dto
-
-enum class Feature(
-    val field : String
-) {
-    TAKE_OUT("takeOut"),
-    ONLY_ONE("onlyOne"),
-    DISASTER("disaster")
-
-}
