+++ src/main/kotlin/mashup/spring/elegant/search/service/SearchService.kt
@@ -1,6 +1,6 @@
 package mashup.spring.elegant.search.service
 
-import mashup.spring.elegant.search.domain.search.SearchType
+import mashup.spring.elegant.search.domain.search.enums.SearchType
 import mashup.spring.elegant.search.dto.SearchDto
 
 interface SearchService {
